Fix #828: Dolphin7.exe fails to start due to ucrtbase.dll error bug
This was caused by an error in a trace message format string. It is on a
code path that would only be followed by the system default locale for
non-Unicode programs is associated with a multi-byte code page. Dolphin
does not support multi-byte ANSI code pages because its ANSI Strings are
fundamentally single-byte character strings. UTF8Strings are preferred for
international languages, with AnsiStrings only really existing for legacy
reasons. Ironically the limitation around multi-byte ACPs also applies to
the CP_UTF8 code page that Windows 10 finally supports (although only in
"beta") as a system code page (it has long been supported as a conversion
source/target by the MultibyteToWideChar/WideCharToMultibyte APIs).

The trace message is fixed, and also some places where the system code page
would be used rather than the one that the VM decides on during startup -
usually these are the same, but not if falling back on 1252 because the
system code page is multi-byte.

Also fixed a bunch of code analysis warnings from the in IDE parser of
VS2019.