Make "Run to Cursor" more conservative
I have found that "Run to Cursor" often runs just *past* the point at which I wanted it to stop.
Because e.g. message arguments are evaluated before the outer message send, clicking near the beginning of a line will often run the entire line and *then* stop, because the cursor is not on top of any of the individual message sends and the includes: check fails. This seems wrong. If we instead just ask "is there any possibility that we are past the cursor", we get much more intuitive behavior, and if the user *did* want to go further, well, that's just a couple "step over"s away...