Add support for binding reference syntax and use for names in Packages
Binding references are a way to provide late binding of fully qualified
names, and are a prerequisite for adding full Namespace support.

- Compiler parses #{Name} into BindingReferences
- Literal BindingReferences have a class context in which they initiate
their search. This is nil for workspace and file-in expressions. Needs to
be included in equality comparison.
- BindingReferences cache the resolved binding (needs further work to
clear this when needed)
- Add meta BindingReferences as this is useful for referencing the class
side, e.g. for method names.
- BindingReferences as names in Packages rather than Symbols
- Initial Refactoring Browser updates for BindingReference use. The RB will
need to be made namespace aware. As a start, replace some obvious
dependencies on the global Smalltalk dictionary and unqualified names.
- Replace direct lookup in Smalltalk from SUnit with use of BindingReference
- Use binding references instead of direct lookups in Smalltalk dictionary
- Compiler supports <mutable> annotation on mutator methods, generating
new primitive inst var setter that ignores object immutability
- Add styling and auto-completion for literal binding references
- Also fix styling/auto-completion of brace array.
- Compile literal strings as UTF-8 by default
- Fix tests, mostly broken by compiling string literals as UTF-8