BindingReference equality was not correct
BindingReferences should only be equal if they contain the exact same path.
It should not be assumed that a ref implicitly rooted in Smalltalk is the
same as one explicitly rooted. BindingReference equality should not be
determined based on whether the refs would resolve to the same binding at
some particular point in time, but on whether they can always be expected
to resolve the same way together in a temporally invariant manner.