Add check to StructureArray>>with:do: that collections are same size
StructureArray>>with:do: may read memory outside the block referenced by
the SA if the 2nd collection is larger. It should be an error to attempt
to perform a with:do: operation with different sized collections.

While here, push up #with:do: to ExternalArray and add a simple unit test.