Add option to use DirectWrite for source code rendering
The DirectWrite option seems to work better in Scintilla than the last
time I tested it, especially the interact with the auto-completion list.
There are still issues in single line mode, so its not yet suitable to
enable by default, but it makes sense to enable it by default for code
workspaces (was already on for comments). There is a new SmalltalkWorkspace
UseDirectDraw option that can be used to turn it off on platforms where
this doesn't work out very well (perhaps Wine).