Fix longstanding bug in Class>>allRoots if Object not naturally first
If Object does not appear first in the collection of roots classes
computed by performing an allInstances scan against Metaclass and
returning the instance classes of those metaclasses that have a nil
superclass, then the Class allRoots method is intended to swap Object
with whatever other class appears first to ensure that Object is always
enumerated first. However the method had a bug where it returned the
result of Array>>at:put:, which is of course the argument (Object).
This causes the boot to fail on some machines with the new parallel all
instances scan that is very likely to return the metaclasses out of order.