Work in progress on #898: Refactor ListControlView>>selectionsByIndex, etc
Only ListView and ListBox subclasses of ListControlView actually support
multiple selections, and of these only ListBox has an irregular API for
querying selections based on whether in single or multi-select mode.
Therefore we can push the complexity of selectionsByIndex down into ListBox
and simplify everything else.
The ListView tests are abstracted to provide coverage for ListBox as well.