Add better default key binding for Duplicate Selection command in Scintilla (#919)
As suggested in https://github.com/dolphinsmalltalk/Dolphin/pull/890#issuecomment-629415585,
provide a default key binding for the Duplicate Line (actually Selection)
command in workspaces that is closer to the VS binding Ctrl+D than the
existing Ctrl+J.

Ctrl+D itself has been bound to 'Display It' in Dolphin workspaces since
the original release in the early 90's, so it would be a bit painful for
muscle memory to take it back now. The suggestion was Ctrl+Alt+D, but aside
from the fact that the Alt key is normally used for accelerators which
have window management or system effect, it is already in use as the key
chord which brings the main Dolphin window to the top of the Z-order.
This leaves Ctrl+Shift+D, which is not currently used.

In order to make it easier to add other default bindings and to reason
about the actual differences in key bindings in ScintillaViews I have
improved the mechanism by which the key bindings are stored so that only
the differences from default are kept as state.