Fix historic hack in STBViewProxy of sending messages to self after become:
STBViewProxy swaps itself with the view it has just restored in
#restoreView:, but then continues and completes the job in its own
message, sending messages that only Views understand to itself. This is
ugly and triggers warnings from the semantic analyser. Although there is
stuff that needs doing after the become:, it should be moved to a method
of View. Also it can be useful for views to know when the rehydration has
finished, so they can perform any fixup they might need, so the operation
should finish with a call to a method (onRestoredFromProxy) that can be
used as an override point.

Also add a View>>onStateRestored override point that is invoked when View
state restoration completes following any of loading from STB, image
restart, or recreation for a creation-time-only style change. This removes
the need for special messages to be added to the end of the View state
just purely for this purpose, of which there were a few cases (ShellView,
Toolbar, and ListView). There is a difference in timing in that the
onStateRestored message is sent after a View and its children have finished
restoring state, rather than after only the View itself has finished
restoring state. However this is a a better time to perform any fixup as
any child views will also be complete.