Revert master to pre-7.2 String>>subStrings: behaviour
This reverts the fix for #877 "String>>subStrings: not symmetric in
treatment of leading/trailing separators" as this was a breaking change
inappropriate for a minor release . The fix will be reinstated in the next
major release (7.2), when the behaviour will also be changed to follow the
ANSI NCITS 319-1998 standard so that it will only accept a String argument,
and will tokenize on any of the characters in the argument String.

Note that all uses in the base system have been replaced by uses of
#split:. The split: implementations in String and , which obeys the correct behaviour that removes the asymmetry of
the historic behaviour whereby it effectively ignored a trailing separator.