Add WriteStream>>grabContents
This is an optimisation for some cases where the number of elements
written to the stream exactly fills the underlying collection, and the
stream is a local of known ownership that is about to go out of scope.
In this case it is possible to grab the streamed over collection as the
contents, rather than copying.