Fix unecessary indirections in IRecordInfo causing SAFEARRAYTest failures
IRecordInfo caller/implementor methods have unnecessary indirections in
handling the array of pointers to BSTRs. By luck this works when the
addresses are in the SmallInteger range because SmallInteger>>yourAddress
returns self. However when the addresses roll over into the LargeInteger
range, this fails because LargeIntegers do not implement `yourAddress`
consistently with SmallInteger. The consistency issue should be fixed, but
in the meantime this makes the failing SAFEARRAYTests always pass.

Contributes to #940.