All Integers should implement #yourAddress consistently
SmallIntegers are immediate objects, so their address cannot be taken. It
is assumed that they are already standing in for addresses, so they
answer self. Although LargeIntegers are not immediate objects (and so
have a body that can be referenced by address), they should behave
consistently with SmallIntegers as otherwise randomly different behaviour
may occur depending on the location in the virtual address space in which
an object is allocated.

Fixes #940.