Ignore class dirtied notifications when PackageManager is not listening
The PackageManager ignores system change events when loading packages to
avoid the overhead of responding to events about code it knows it is
loading. The "isDirty" notifications sent by a class to set the change
flag/position on a package were side-stepping this mechanism by going
directly to the package, rather than through the PackageManager.

Also, CompiledMethod>>infoTip should be in development package