Fix ProcessTest>>testLargeAddress to work when MEM_TOP_DOWN is enabled
Windows has a debugging mode for 32-bit processes that starts allocating
memory from the top of the virtual address space, rather than the bottom.
This will cause most object addresses to be > 2Gb, and also outside the
SmallInteger range. ProcessTest>>testLargeAddress is designed to test
what happens with Process copying when the Process objects are allocated
above the 2Gb boundary, and includes some code to reserve sufficient memory
to push allocations over that boundary. When running in MEM_TOP_DOWN this
pre-allocation is likely to do the reverse.