Define SequenceableCollection>>copyWith:atIndex: and copyWithoutIndex:
Extract the generally useful code added recently for inserting an element
into the middle of an Array from Behavior>>addSubclass: into a new
`copyWith:atIndex:` method of SequenceableCollection, with test.

Add a complementary #copyWithoutIndex: that could be used for the subclass
removal and similar operations to more efficiently remove a single entry
from an Array by index.