Add package class/inst properties to govern class pre-declaration on load
Forward declaration of classes has a significant effect on package load
times, but is historically what Dolphin has done. I recently disabled this
unless packages contained source globals (e.g. pool constants dictionaries)
but the effect is to break loading of code ported from Pharo that uses
the idiom of representing pool dictionaries as classes. Classes as pools
may be referenced before they are defined, and the hierarchy ordering
of classes does not help. In order to re-enable scenarios that worked
previously, but which now do not, but still provide options to disable
pre-declaration when it is known not to be required, this commit adds
properties on the class and instance sides of Package. The class side
setting controls the default choice when there is no per-package setting.
The default for this default is true, to behave as before (this may be
changed in future). The setting can also be applied to individual packages
and once set is persisted in the .pax file and respected regardless of
the overall setting.