Restore COM extended error info reporting for failed COM calls
This was lost 9ffb339942f86cc8a37b060dac0dc0e581f7d51e

It would be nice to add a test for this, but first I'll have to identify
a suitable friendly COM component with which sets IErrorInfo to provoke
into providing an error. The only COM object implemented in the image which
does this is AXPropertyBag, and it is returning a list of errors through
a special log parameter.