Fix #967: ListView checkboxes mode not working
The ListView checkboxes mode has never been properly supported in the MVP
system, but the option to set the mode was exposed. The feature relies on
using the state image list (and actually it isn't much more than that,
plus reacting to clicking over the image to toggle between images), and
was broken by the recent pulling-up of some state image management that
had only been implemented in the ListTreeView subclass (see
b4e59a2a7ba482b683db8a2e53ca0ef012c7df74).
This fix restores the fundamental operation of checkboxes and adds get/set
methods to make working with the feature a little easier, but it still
doesn't constitute a fully supported feature and there are a number of
limitations and inadequacies:
- Doesn't work in the default virtual mode
- View feature only - no support in Presenter/Model layers, and no events
to observe.
- Mutually exclusive with other uses of state images such as in ListTreeView