Improve handling of unrecoverable out of memory conditions
When the virtual memory space is exhausted, continuing fairly quickly
results in unpredictable failures that mask the underlying problem of
runnig out of memory.

Normally when the STATUS_NO_MEMORY exception is raised, the VM sends an
interrupt into the image so that it can attempt to recover. This avoids
prematurely terminating the app when the error condition is transient, for
example attempt to allocate a very large object due to a logic error.
However, sometimes the error will be raised when there really is
insufficient remaining 32-bit address space to meet the allocation request.
Normally this will be when approach 4Gb of space has been been allocated
already (perhaps due to a creating a very large number of Process objects
that are either live, or which are referenced from finalizable garbage
that is not being finalized due to a loop running at higher priority than
finalization). The next attempt to allocate a further process will then
fail, or if the finalization queue has become very large the attempt to
extend it will fail, or normal object allocation may fail because the
heap cannot be extended. In practice this seems to happen when there is
less than 100Mb of free address space, although due to fragmentation the
maximum contiguous space may be substantially smaller than this. When this
occurs, continuing to execute Smalltalk code will quickly result in a
stack overflow caused by the Smalltalk code attempting to respond to the
no-memory interrupt, having to make an allocation to do so, and generating
a further no-memory interrupt. This continues for a while, until the stack
blows. Stack overflows may result in the OS fast terminating the process
with no error diagnostic - it just disappears - because there is
insufficient space on the main thread stack to do any error reporting.

With this change when an allocation fails the VM will first check the
amount of virtual memory available, and if it is below a threshold it
will trigger a crash dump and fatal application exit with an a diagnostic
message describing the out of memory condition.

Contributes to #960