Fix for #960
Fix race condition where overlapped calls could be terminated before they had started
to make the call, but still continue and make the call after the Smalltalk stack had
been torn down. This caused an difficult to debug condition where the object table
was being overwritten by the called API, resulting in a later failure in GC activities.

Also includes a lot of work to improve the robustness of error handling in overlapped
calls so that access violiations, FP faults, out of memory conditions, etc, are all
handled, and where appropriate re-raised on the main thread. These are all now tested
in VMTest.

The VM now reports individual STATUS_INVALID_PARAMETER_<n> failure codes for
errors on parameters 1 to 12, and then a generic STATUS_INVALID_PARAMETER
for parameters 13+. There are currently only 4 external methods with 13 or more parameters in the standard image out of 3065 (0.13%): 2 are in GDI+, one is
StretchDIBits, and the other is in the ODBCLibrary. Testing of this is improved and
extended to cover overlapped calls too.

Improve error reporting added in e08e29826a0635caed5c7a0bdb9c929c8fcfddfb when
unrecoverable out of (virtual) memory condition occurs.