Increase range of classes that can be printed as constants
Add #isSelfEvaluating and related changes to allow for a much wider
range of constant definitions than just literals. Points are an
example.

Arrays can now print in literal, brace array, or non-evaluable forms.

LimitedWriteStream added. This can be used to wrap another writeable
stream and impose a limit on the amount that can be written to it.
This replaces the various custom implementations of this in printOn:
methods such as that for Collection. It also maintains a visited
set to detect recursive printing on an opt in basis.

Remove Process._alreadyPrinted inst var as no longer used

Also fix tests broken by Object>>printString refactoring