Replace loose method name cache with 2-level store
With the extra name complexity caused by namespaces, the old single level
store no longer performs adequately. If we partition the cache by class we
get improved perf for a number of scenarios that need to find all the
loose methods in a class, and it also means we can maintain the cache
more easily and so rebuild it less often.