Fix registry serialization (needs to use full class names now)
Backwards compatibility is possible, as long as the old global names remain
bindable. This will be the case for BCL classes in Core, Kernel, etc. For
other classes an alias may be required.

Forwards compatibility is not supported, so to avoid breaking old versions
of an application that uses registry serialisation, a new registry key
should be used.

XmlPad is adapted to illustrate maintaining compatibility (loads old
settings, and stores new settings under a different key)