Change system folder background colour to distinguish 7.2 easily
Change the background colour of the system folder and splash screen to
help distinguish a 7.2 image from 7.1, which can be helpful when working
with both. This is not necessarily permanent.

The splash screen background is changed using a color key substitution.
The results of this are not perfect, e.g. around smoothed edges bordering
blocks of other colours, but it is good enough for now.