Selecting from SortedCollection performs unecessary re-sort
The #select: implementation inherited by SortedCollection works, but ends
up (in effect) sorting the selected elements again, which is unnecessary
since we know they will be selected in sorted order as they are enumerated.
This speeds up select: operations on SortedCollections by around an order
of magnitude, depending on the sort comparison block overhead.