Improvements related to Float printing issues discussed in #1026
- Float>>printOn:decimalPlaces: is slow. Not addressed directly, but
rather the alternative of using Locale Float printing is suggested. This is
faster because it uses the inexact Float printing to generate the
intermediate result that is then formatted to a number of decimal places
by the Windows GetNumberFormat API.
- A way to print with comma (or dot) separated number groups is useful for
displaying large numbers in a more readable way in applications - the
aforementioned Float printing through Locale and GetNumberFormat can do
this, among other things.
- Make it a bit easier to call the Locale Float printing, and make a few
small performance tweaks.
- Add a few extra cases to the Locale Float printing tests, and fix a few
rounding errors at the limits of precision caused by not using the full
available precision in the intermediate value.
- Make it possible to change the default boundaries for when Floats are
printed in exponential form by changing the value of a class variable
rather than changing the code (see Float.DefaultDecimalExponents)