Fix #1042: OneOffNoteBubble sends unimplemented message in deployed apps
#productRegistryKey was sent to the current SessionManager by
OneOffNoteBubble to get the name of the registry key under which it stores
a record of the "one-off" notes that have previously been shown. Only the
DevelopmentSessionManager had an implementation of productRegistryKey, so
this would cause failures in a typical deployed application that uses
OneOffNoteBubble. To fix, added a generic definition of productRegistryKey
in SessionManager.

Also add a PackageClosureTest for EditableListView demo to help catch
broken deployment dependencies, and also because his test showed up issue
#1042, and so acts as a regression test for the issue.