Put literals not used in code at end of literal frame
Some literals are retained for reference searching purposes, but are not
used in the actual code, e.g. for optimised control selectors such as
`whileTrue:` or the literals of ##() expressions. These should be placed
at the end of the literal frame to avoid the literal indices in code
being larger than they need to be (which can lengthen the instructions).
The effect of this is generally marginal, but for some longer methods it
makes a difference.