Partition base system into Namespaces: Part 5, MVP/Deprecated package
The deprecated classes are left in the Smalltalk namespace.
Aliases of old names in Smalltalk are added to redirect to the namespaced
class. This helps with loading and fixing up older code, although it is
only necessary when a class has been renamed as the UI, Graphics, etc,
namespaces are imported into the legacy Smalltalk namespace so base
classes with unqualified names that have not been changed will still be
visible in code compiled in the Smalltalk namespace. All pre-existing code
is loaded into the Smalltalk namespace by default.

Not bootable