Partition base system into Namespaces: Part 2b, Dolphin/Base packages
The Namespace class is now defined, providing the abstract definition of
namespaces, although most of the namespace behavior is actually
implemented on Class. The Dolphin namespace system is unusual in being
class based - i.e. the actual namespaces are sub-classes of Namespace, not
instances of Namespace. This has a number of important advantages over
the instance-based namespaces one generally finds in other Smalltalk
systems:
- Reduction in cognitive complexity. Class-based namespaces are just
an extension of existing concepts, rather than being a completely new
concept.
- Uniformity - classes are namespace, and namespaces are classes.
- Nested classes are supported naturally.
- Existing tooling requires less adaptation - for example creating a new
Namespace is already supported by simply defining a new subclass. Namespaces
do not require much adaptation of the browsers.
- Simpler implementation: There is only a small logical distinction between
class variables and namespace variables, and physically they have the same
implementation as members of the classPool of Namespace classes. The
logical distinction is that namespace variables are not considered to be
owned by the (Namespace) class whose pool they live in, whereas class
variables are owned by that class. Unlike class variables, the addition
and removal of namespace variables is not considered to be a change to
the (Namespace) class.
- Behaviour can be easily and naturally defined on specific namespaces.

As Namespaces are classes, their behaviour is defined on their own class
side, and (largely) inherited from the instance side of Class.
Dictionary-like behaviour is provided for convenience on Namespace class as
well as some convenience enumerations, like #allBehaviors. Binding
resolution is mainly implemented on Class, although Namespace adds a subtle
variation to this to avoid searching its own hierarchy twice. This is
because it turned out to make sense to define sub-namespaces as sub-classes
of their enclosing namespace. This arrangement this fits more naturally in
the tooling (and probably in the developers' mental model too). It is
really just a convenient convention though, and is not necessary as there
need not be a sub-type relationship. The more important relationship
between namespaces is through the 'environment' class instance variable,
through which all Class objects refer to their enclosing scope.

As mentioned above, since Dolphin namespaces are classes, and Class
defines basic namespace behavior, nested classes are possible within the
model. At present there is no tooling support for nested classes, however.

The basic namespaces has been chosen for compatibility with existing
precedent:
  Root (the root namespace)
  Smalltalk (the system and default namespace for backwards compat)
  Core - home for most of the public base class library for use from apps
  Kernel - system implementation classes, less likely to be used in apps
  OS - platform-specific (Windows) FFI
  External - non-platform specific FFI.
  Graphics - graphical elements other than Views such as Bitmaps, Icons
  UI - MVP GUI classes.

All of these namespaces are defined in the base package, including UI and
Graphics, which logically are part of MVP:
- UI is required for MessageBox and ResourceIdentifier.
- Graphics defined in the base so it can be imported into Smalltalk for
backwards compatibility, although it has no members in the base system
until MVP is loaded.

Note that the basic namespaces are all defined in Root, not Smalltalk.
This choice has been made to avoid polluting the new namespaces with the
entire contents of Smalltalk (since names in enclosing scopes are
implicitly visible). Smalltalk itself is present to support legacy code,
and because it is convenient as a playground. It imports all the base
namespaces so the effect is that most classes are visible to code defined
in it using just unqualified names. However the reverse is not true -
classes defined in Smalltalk will not be visible from Core, for example.
This can cause some confusion at first for anyone used to working in a
single global namespace.

Note that at present UI is not imported into Smalltalk, although this may
be changed.

Not bootable - all the commits in the "Partion base system" sequence
must be included to reach a point where the system is bootable and testable.