Partition base system into Namespaces: Part 67b, Camp Smalltalk contrib
All but SUnit not namespaced, but some changes required due to introduction
of base system namespaces.

SUnit is renamed into XProgramming.SUnit namespace, including the OA
extensions.

A new SUnitLegacy package is created that defines aliases for all the
original unqualified names in the Smalltalk namespace. This is to avoid
having to import XProgramming.SUnit into Smalltalk.