Temporary fix for regression in ListView #1071
The change to resize the columns only when the area is changed rather than when the
rectangle is changed has caused the ListView's scroll origin to become corrupted so
that it does not display correctly when resetting to a smaller number of items.

This resolution is not ideal, but the issue impacts on normal use to leave unmitigated.