Fix #1070 (or at least prevent the confusion resulting in #1070)
The presence of an obsolete default in FetchVM.ps1 for the VM version to
fetch is not helpful. Even less helpful is the combination of this with
out of date instructions in the readme that suggest running it directly,
rather than the wrapper script FetchVM.cmd. It is the latter that
determines the correct version to download.