Fix #1078: Add vc_redist to setup
Oddly there seems to be no "official" way to do this on the InnoSetup
website. The only relevant KB article is for VC++6.0!, and recommends
cracking open the .exe and installing all the individual DLLs, which is
not a supported way to distribute the VC runtime. There are various
stack overflow and other suggestions online, but all of these seem
overcomplicated and will required maintenance of version checks.

So, the KISS approach:
- Include the redist from the VC installation on the build machine
(ensures compatibility)
- Don't bother 2nd guessing the already-installed check of the redist setup
- Run the redist setup after file installation