Speed up rebuild of SmalltalkSystem message selector cache (#1088)
* Speed up rebuild of SmalltalkSystem message selector cache

SmalltalkSystem maintains a cache of all implemented messages in the system
that is used (e.g.) for auto-completion. The cache is reset when either a
new method is added, or removed, or a class is removed. Rebuilding the
cache is generally fast (sub-100ms), but might take long enough to cause
a noticeable delay in a very large image with a lot of classes and methods.
We can easily speed up the the underlying Behaviour>>understoodSelectorsDo:
method and about halve the time the cache rebuild takes.

* Behavior>>allSelectors and canUnderstand: should include dynamic selectors

* Fix a couple of sends of unimplemented messages