Fix some basic issues with application deployment
This allows for deployment of some MVP samples such as Notepad, XmlPad,
Calculator, Etch-a-Sketch. The latter fails to start, however, reporting
a Gdiplus error.

Deployment of console applications still fails late in deployment due to
an infinite recursion attempting to resolve an import into Smalltalk after
the UI namespace has been removed.

- Logging of class names should report full class names
- Namespaces being removed as not included among class references (a Class
implicitly references its environment).
- Nil'ing out of classPools is destroying the Namespaces, causing
deployment to fail attempting to look up BindingReferences
- Mixup of bindings/names in BasicImageStripper>>#emptySystemDictionary
- Many unbound classes remain because of unflushed BindingReferences holding
onto them through cached bindings
- Class imports must be set to empty Array, not nil
- Dependency of "Dolphin (Old Names)" on External.Field, through the
ExternalField alias, causes an uncommitted pre-requisite to be reported
if any of the application packages depend on Old Names, which is fatal.
- Scintilla instances left open (or opened) during deployment will cause
AV's when the ScintillaLibrary default instance is closed, as it is not
lazily re-opened.
- Uninstallation of certain development classes during deployment where
the application depends on the development system package(s) (so they are not
removed as part of the uninstallation of dev sys packages), can cause a
new SmalltalkSystem instance to be created, and to register for change
events, causing notifications to begin being propagated to Packages, etc,
which will fail as the package system has been at least partially removed.