Fix error on startup of GUI apps with InternalIcons
The Gdiplus startup 'Invalid Parameter' issue is a symptom of the wider
problem that GraphicsTools saved in the image are not being re-initialized
on startup. This was because UI.GuiInputState refers to various classes
that require startup using unqualified binding references. These bind
successfully in normal circumstances because UI imports the Graphics
namespace. However the image stripper was clearning out all namespace
imports, breaking the ability to resolve such unqualified BindingReferences
It is probably unsafe to remove class imports, as it will affect such
symbolic references and may prevent them resolving in the same way they
did in the development image. Although fully qualifying the references
would also fix the problem, that rather defeats allowing unqualified or
partially qualified references that are bound at runtime in context.