Fix #1103: SocketAbstract class>>#onShutdown improperly stripped
The SocketAbstract class initialize method is forcibly retained in order
to ensure there is a referenced to the onStartup/onShutdown methods.

I did try and enhance the PackageClosureTests to be able to detect the
original error, but was defeated by self-refential send that makes the
message part of the BrowserEnvironment call graph, but which the
application deployment process will specifically ignore. This would be
complicated (and slow) to simulate, and given that the package with the
fault is obsolete, I decided it wasn't worth further effort this time.

I retained the modifications to the PackageClosureTests because it does
make the simulation slightly more accurate. However it will slow them
down a bit.