Fix some package scripts that fail to compile
When the 'Dolphin Anonymous Classes' package is unloaded, the package
scripts are compiled in the Kernel namespace, so may not have visibility
of classes in the package. Creating an anonymous class in order to be
able to provide a binding context that includes the packages instance
variables and which imports Smalltalk, is a bit too much of a clever trick.
Really what is needed is a way to pass additional binding context to the
Compiler that can be used to compile scripts. For now, just fix the scripts
by using qualified class names.

Also fix damage done when Burning River/KernelLibraryExtensions is unloaded
as it is removing a constant that is now part of the base.