Fix broken add instance variable refactoring in browser UI
Using the UI to add an instance variable appears to work, but then one
can't actually refer to the new instance variable name in code. This is
because the instance variable name gets added to the class' collection as
a Utf16String, and the Compiler is only able to interpret Utf8Strings.