Remove/deprecate some ambiguous base system messages
The use of non-specific messages in the base system for a number of
separate purposes is undesirable for a number of reasons, as in a
dynamically typed system one cannot statically distinguish that the
implementations of the message are for separate purposes. This gets in
the way of effective reference/definitions searches, can be a problem for
correctly scoping refactorings, and also impacts image stripping
effectiveness.