Workaround fix for infinite loop in deleted Smaltalk imports lookup
When one of the imports into Smalltalk is deleted, subsequent attempts
to resolve a binding using Smalltalk when that binding is not in Smalltalk
will cause an infinite loop attempting to traverse Smalltalk's imports.
This is a bug in the binding resolution, but needs some thought to fix.
As a temporary solution, which is in any case tidy, ensure that removal
of the UI or Graphics namespaces removes the associated import into
Smalltalk.