Rename Dolphin Sockets classes to remove '2' suffix
Now that the old Sockets package is back in Smalltalk, and the new
Sockets package is in the Net namespace, there is no need for a namespace
suffix as the classes are in different namespaces. Although this does
make the classes ambiguous from a human perspective, the expectation is
that the old sockets package will no longer be part of the default image.

Also relocate the Chat sample to a new Net.Examples namespace.