Remove some unnecessary/incorrect CRT dependencies
Every function in the CRT that is used from the image has to be built
into the to-go application stubs, whether used or not. Remove/replace
some that are unused or have Windows API equivalents.

- _strrev does not work correctly for AnsiString if the string contains
embedded nulls.
- Some non-wide string CRT functions that operate on paths were being
used (_chmod, _chdir) and so did not work correctly for non-ASCII file
paths.
- Character class testing methods do not work correctly for many Unicode
code points. This is improved, but not completely fixed as the underlying
Windows API does not provide correct results for codepoints that require
more than one UTF-16 code unit.
API
- Clear out all the pre-existing deprected CRTLibrary methods.